refactor: 모듈화 및 리팩토링
