+++ .prettierrc
@@ -1,7 +1,6 @@
 {
   "plugins": ["@trivago/prettier-plugin-sort-imports"],
   "importOrder": [
-    "^node:.*$",
     "<THIRD_PARTY_MODULES>",
     "^@/(apis|assets|constants|hooks|mocks|pages|stores|utils)/(.*)$",
     "^@/components/(.*)$",
