+++ src/reset.css
@@ -12,10 +12,6 @@
   border-width: 0;
   border-style: solid;
 }
-::before,
-::after {
-  content: '';
-}
 html {
   line-height: 1.5;
   tab-size: 4;
