+++ src/stories/Grid.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { v4 as uuidv4 } from 'uuid';
 
-import { Grid } from '@/components/ui/Grid';
+import { Grid } from '@/components/ui/Layout/Grid';
 
 const NumberNode = () => {
   return (
@@ -14,8 +14,9 @@ const NumberNode = () => {
 };
 
 const meta: Meta<typeof Grid> = {
-  title: 'Example/Grid',
+  title: 'Example/Layout/Grid',
   component: Grid,
+  tags: ['autodocs'],
   parameters: {
     layout: 'centered',
   },
@@ -38,6 +39,18 @@ type Story = StoryObj<typeof Grid>;
 export const Default: Story = {
   args: {
     gap: 10,
-    columns: 4,
+    columns: 3,
+  },
+};
+
+export const ResponsiveColumns: Story = {
+  args: {
+    gap: 10,
+    columns: {
+      initial: 1,
+      lg: 4,
+      md: 3,
+      sm: 2,
+    },
   },
 };
