+++ src/styles/theme.ts
@@ -1,0 +1,16 @@
+export const colors = {
+  yellow: {
+    100: '#FFF9C4',
+    200: '#FFF59D',
+    500: '#FFEB3B',
+  },
+  gray: {
+    100: '#F5F5F5',
+    200: '#EEEEEE',
+    300: '#E0E0E0',
+    400: '#BDBDBD',
+    500: '#9E9E9E',
+    600: '#757575',
+    800: '#424242',
+  },
+};
